# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClientConfig {
  gameServerUrl: string;
  fromEnv: boolean;
}

/**
 * Everything the server needs to act for this player. The browser owns them and sends them with every call, so no request depends on state left behind by an earlier one.
 */
export interface PlayerCredentials {
  /**
     * @minLength 1
     * @maxLength 64
     */
  token: string;
  /**
     * @minLength 1
     * @maxLength 24
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  playerKey: string;
  /**
     * @minLength 1
     * @maxLength 2048
     */
  gameServerUrl: string;
}

export interface Player {
  playerId: string;
  displayName: string;
  mode: string;
  deployed: boolean;
}

export interface JoinResponse {
  ok: boolean;
  player?: Player;
}

export interface RoundPart {
  kind: string;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  index?: number | null;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  word?: string | null;
  /** @nullable */
  by?: string | null;
  /** @nullable */
  filler?: boolean | null;
}

export interface Round {
  roundId: string;
  seq: number;
  phase: string;
  title: string;
  /** @nullable */
  endsAt: number | null;
  blankCount: number;
  parts: RoundPart[];
}

export interface YourBlank {
  /** @nullable */
  blankIndex: number | null;
  /** @nullable */
  hint: string | null;
  /** @nullable */
  example: string | null;
  /** @nullable */
  word: string | null;
  submitted: boolean;
}

export interface Stats {
  players: number;
  active: number;
  deployed: number;
  local: number;
}

export interface PollResponse {
  ok: boolean;
  now?: number;
  player?: Player;
  round?: Round | null;
  you?: YourBlank;
  stats?: Stats;
  error?: string;
}

export interface SubmitInput {
  /**
     * @minLength 1
     * @maxLength 64
     */
  token: string;
  /**
     * @minLength 1
     * @maxLength 24
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  playerKey: string;
  /**
     * @minLength 1
     * @maxLength 2048
     */
  gameServerUrl: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  word: string;
}

export interface SubmitResponse {
  ok: boolean;
  word?: string;
  blankIndex?: number;
  error?: string;
}

export interface ErrorResponse {
  ok: boolean;
  error: string;
}