# lib/api-zod/src/generated/types/joinResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface JoinResponse {
  ok: boolean;
  player?: Player;
}