# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  playerId: string;
  displayName: string;
  mode: string;
  deployed: boolean;
}