# lib/api-zod/src/generated/types/playerCredentials.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Everything the server needs to act for this player. The browser owns them and sends them with every call, so no request depends on state left behind by an earlier one.
 */
export interface PlayerCredentials {
  /**
     * @minLength 1
     * @maxLength 64
     */
  token: string;
  /**
     * @minLength 1
     * @maxLength 24
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  playerKey: string;
  /**
     * @minLength 1
     * @maxLength 2048
     */
  gameServerUrl: string;
}