# lib/api-zod/src/generated/types/pollResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';
import type { Round } from './round';
import type { Stats } from './stats';
import type { YourBlank } from './yourBlank';

export interface PollResponse {
  ok: boolean;
  now?: number;
  player?: Player;
  round?: Round | null;
  you?: YourBlank;
  stats?: Stats;
  error?: string;
}