# lib/api-zod/src/generated/types/round.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoundPart } from './roundPart';

export interface Round {
  roundId: string;
  seq: number;
  phase: string;
  title: string;
  /** @nullable */
  endsAt: number | null;
  blankCount: number;
  parts: RoundPart[];
}