# lib/api-zod/src/generated/types/roundPart.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoundPart {
  kind: string;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  index?: number | null;
  /** @nullable */
  hint?: string | null;
  /** @nullable */
  example?: string | null;
  /** @nullable */
  word?: string | null;
  /** @nullable */
  by?: string | null;
  /** @nullable */
  filler?: boolean | null;
}