# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  players: number;
  active: number;
  deployed: number;
  local: number;
}