# lib/api-zod/src/generated/types/submitInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitInput {
  /**
     * @minLength 1
     * @maxLength 64
     */
  token: string;
  /**
     * @minLength 1
     * @maxLength 24
     */
  displayName: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  playerKey: string;
  /**
     * @minLength 1
     * @maxLength 2048
     */
  gameServerUrl: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  word: string;
}