# lib/api-zod/src/generated/types/submitResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResponse {
  ok: boolean;
  word?: string;
  blankIndex?: number;
  error?: string;
}