# lib/api-zod/src/generated/types/yourBlank.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface YourBlank {
  /** @nullable */
  blankIndex: number | null;
  /** @nullable */
  hint: string | null;
  /** @nullable */
  example: string | null;
  /** @nullable */
  word: string | null;
  submitted: boolean;
}